requirements.txt
django-import-export
django-ckeditor
patterns
